@@ -24,17 +24,18 @@
     "browserify": "^14.4.0",
     "buffer-shims": "^1.0.0",
     "cross-env": "^5.0.0",
-    "eslint": "3.19.0",
-    "eslint-plugin-node": "5.0.0",
-    "eslint-plugin-security": "1.3.0",
-    "htmllint-cli": "0.0.6",
+    "eslint": "^4.0.0",
+    "eslint-plugin-mocha": "^4.11.0",
+    "eslint-plugin-node": "^5.0.0",
+    "eslint-plugin-security": "^1.4.0",
+    "htmllint-cli": "^0.0.6",
     "mocha": "^3.4.2",
-    "npm-run-all": "4.0.2",
-    "prettier": "1.4.4",
+    "npm-run-all": "^4.0.2",
+    "prettier": "^1.4.4",
     "proxyquire": "^1.8.0",
     "sinon": "^2.3.5",
-    "stylelint": "7.11.0",
-    "stylelint-config-standard": "16.0.0",
+    "stylelint": "^7.11.0",
+    "stylelint-config-standard": "^16.0.0",
     "watchify": "^3.9.0"
   },
   "engines": {
@@ -44,7 +45,7 @@
   "repository": "mozilla/something-awesome",
   "scripts": {
     "dev": "watchify frontend/src/main.js -o public/bundle.js -d | node server/portal_server.js",
-    "format": "prettier 'frontend/src/*.js' 'server/*.js' 'public/*.css' --single-quote --write",
+    "format": "prettier 'frontend/src/*.js' 'public/*.css' 'server/*.js' 'test/*.js' --single-quote --write",
     "lint": "npm-run-all lint:*",
     "lint:css": "stylelint 'public/*.css'",
     "lint:html": "htmllint 'views/*.handlebars'",
